Checksum and minor fixes (#119)
* fix checksum in beatles and salami, simplify tar.gz downloader

* beat positions to int rwc

* fix download folder

* remove comments

* update version

+++ mirdata/rwc_classical.py
@@ -252,7 +252,7 @@ def _load_beats(beats_path):
         np.array(beat_positions), np.array(beat_times)
     )
 
-    return utils.BeatData(beat_times, beat_positions)
+    return utils.BeatData(beat_times, beat_positions.astype(int))
 
 
 def _load_metadata(data_home):
